Add robust self location technique to geant4.csh
The use of $_ is not robust as this variable is empty if the script is
sourced inside another script (e.g. .cshrc).

We check for the emptiness of $_, and if it is we see if the script was
either sourced from the directory in which it exists, or this directory
was supplied as an argument to the script.

If we are not able to locate it after this, we print an error message to
stdout. We don't return a full error, but no further operations are
performed, so the environment will not be modified.

Some basic tests are implemented to confirm the error output, and that
the two alternate location methods work.