Add system/compiler detection from Geant4 9.5
A key feature of the Geant4 GNUmake system is the G4SYSTEM environment
variable which should be set to "SYSTEMNAME-COMPILERNAME". The two
text entries depend on the OS, the latter on the C++ compiler used.

Port in the detection method of these quantities from the current
Geant4 mainline. This is basic, and does not detect the Clang compiler
corrected on CMake <= 2.8.1. The clang detection should be present in
CMake 2.8.2 and higher, so we need to backport the detection.