Implement factory for G4UIsession instances
Implement UISessionFactory class for creating concrete G4UIsession
instances. The factory maps string ids to factory functions, one
per session. If a session cannot be created, the default policy
is to return a null pointer.

Add helper function BuildUISessionFactory to return a factory instance
populated with all sessions available in the Geant4 build we are
using.

Use helper function and resultant factory instance to create a
session instance (qt only) immediately on app startup. This allows
all logging in the app to go direct to the session. In GUI sessions,
output can be outputted to terminal unless this is done. Adjust
instantiation of G4VisExecutive and configuration macro application
so that Qt session is fully primed with an OGL viewer on session start.