Store new'd resources in boost::shared_ptr
Use Boost's shared_ptr to hold the UI session, run manager and vis
manager instances. Remove the delete commands at the end of the
application as the shared pointers will garbage collect them.

Find the Boost package at the top level. Only find the headers, as
shared_ptr is header only. Ensure that the Boost include path is
added so that they are found.