Implement CommandLineParser using Boost program_options
Add help and session arguments, the latter holding a string which
can be passed to the UI session factory. If an unrecognized option
is found, print help and exit with failure. If help option is
found, print help and exit. Parse command line by forwarding
argc and argv to parser implementation held by CommandLineParser.

Remove check for batch mode by unit argc in main program. Functionality
for running macro via positional argument not yet added in.

Aim of this commit is to demonstrate basic functionality of
a Geant4 app with command line selection of UI session.