Update guide to cover new command line interface
Note dependency on Boost, and add description of how to get CMake to
locate Boost if it is not in a system/CMake known location.

Show usage of command line arguments, especially -h/--help.

Change interactive mode description to describe choice of default
(G)UI. Add note that batch mode does not yet support supplying a
macro.

Update User Interfaces section with description of UISessionFactory
class and the session available.