Implement GMake function(s) for defining datasets
Provide Geant4GMake.gmk.in with a callable function to define a Geant4
dataset. When called with the dataset (environment variable) name and
default path, check for definition of name. If the name is defined,
simply append it to the list of known names. This ensures the user
can still define the datasets via the environment, command line or
other makefile. If the name is not defined, append it the the list of
names, and set its value to the supplied default path.

For now, just define the G4LEDATA variable with a dummy value for
checking.

Add a geant4_datasets target in the targets file to report the dataset
names, where their value comes from, and the value itself.