Demonstrate querying of user setting of options
Geant4GMake sets "option variables" to indicate whether the install
provides a specific feature, e.g. GDML. The user should be able to
set, via environment, make or command line, a variable to request
activation of a feature. If the install does not support the feature,
the build should fail.

Ensure all option variables get added to the list of allowed options,
whether they are supported by the install or not. Iterate over list
of options as last task in Geant4GMake, seeing if the user has
requested activation of the feature. Log to stdout the requested
options. In a second iteration, fail if user has requested a feature
not supported by the current install.